extracted interface + log for parserbatchcontroller
