+++ service/src/main/java/cz/fi/muni/pv243/service/batching/ParserBatchControllerImpl.java
@@ -3,12 +3,15 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package cz.fi.muni.pv243.service.batching.controller;
+package cz.fi.muni.pv243.service.batching;
 
 import java.util.Properties;
 import javax.batch.operations.JobOperator;
 import javax.batch.runtime.BatchRuntime;
+import javax.ejb.Schedule;
 import javax.enterprise.context.ApplicationScoped;
+
+import cz.fi.muni.pv243.service.logging.ParserManagerLogger;
 
 /**
  *
@@ -19,12 +22,16 @@ public class ParserBatchControllerImpl implements ParserBatchController {
 
     private final String parserJobName = "parser-job";
     
+    /* (non-Javadoc)
+     * @see cz.fi.muni.pv243.service.batching.ParserBatchController#startJob()
+     */
     @Override
+    @Schedule(dayOfWeek = "Mon", hour = "0"/*, persistent = "false"*/)
     public void startJob() {
         
         JobOperator jobOperator = BatchRuntime.getJobOperator();
         Properties jobParameters = new Properties();
         long execID = jobOperator.start(parserJobName, jobParameters);
-        //log
+        ParserManagerLogger.LOGGER.logBatchJobRunning(execID);
     }
 }
