--- service/src/main/java/cz/fi/muni/pv243/service/batching/controller/ParserBatchController.java
@@ -1,15 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package cz.fi.muni.pv243.service.batching.controller;
-
-/**
- *
- * @author Michaela Bocanova
- */
-public interface ParserBatchController {
-    
-    public void startJob();
-}
